+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 # v0.7.0:
 * Added the Generalized Alignment Index method: `gali`
 * Added GALI for continuous systems.
-* Added the Henon Helies system in famous_systems. TODO is still to add
-  callbacks that conserve energy.
+* Added the Henon Helies system in famous_systems.
+* Change the pretty-printing of all systems, by adding an additional
+  `name` field everywhere that can be used for convenience.
 * Bugfix of the printing of `Dataset` on Juno.
 * Added `BigDiscreteDS` (still need methods for e.g. lyapunovs).
   * Still need to add docs for it.
+* Added automated jacobian computation for `BigDiscreteDS`.
 
 # v0.6.0:
 * Many minor bugfixes and many corrections to the documentation examples.